+++ crates/but/src/command/legacy/rub/mod.rs
@@ -318,14 +318,27 @@ impl<'a> UncommittedToBranchOperation<'a> {
 impl<'a> UncommittedToStackOperation<'a> {
     /// Executes this operation.
     pub(crate) fn execute(self, ctx: &mut Context, out: &mut OutputChannel) -> anyhow::Result<()> {
-        create_snapshot(ctx, OperationKind::MoveHunk);
-        assign::assign_uncommitted_to_stack(
-            ctx,
-            self.hunk_assignments,
-            self.description,
-            &self.stack_id,
-            out,
-        )
+        self.execute_inner(ctx)?;
+        if let Some(out) = out.for_human() {
+            writeln!(
+                out,
+                "Staged {} → stack {}.",
+                self.description,
+                format!("[{}]", self.stack_id).green()
+            )?;
+        } else if let Some(out) = out.for_json() {
+            out.write_value(serde_json::json!({"ok": true}))?;
+        }
+        Ok(())
+    }
+
+    /// Executes `UncommittedToStack` by assigning selected hunks to the target stack.
+    pub(crate) fn execute_inner(&self, ctx: &mut Context) -> anyhow::Result<()> {
+        let requests = assignment_requests_for_selected_hunks(
+            self.hunk_assignments.iter().copied(),
+            Some(self.stack_id),
+        );
+        but_api::diff::assign_hunk(ctx, requests)
     }
 }
